Move PersistenceSender
